Moisha_Back/requirements.txt
Django==2.1.1
django-cors-headers==2.4.0
django-haystack==2.8.1
djangorestframework==3.9.0
elasticsearch==2.4.1
pytz==2018.5
urllib3==1.13.1

